+++ packages/openapi/src/ui/operation/usage-tabs/index.tsx
@@ -1,6 +1,9 @@
 import type { MethodInformation, RenderContext } from '@/types';
-import type { SampleGenerator } from '@/requests/types';
-import { defaultSamples } from '@/requests/generators';
+import type { CodeUsageGeneratorFn } from '@/requests/generators';
+import {
+  type CodeUsageGeneratorRegistry,
+  createCodeUsageGeneratorRegistry,
+} from '@/requests/generators';
 import {
   CodeBlockTab,
   CodeBlockTabs,
@@ -9,6 +12,7 @@ import {
 } from 'fumadocs-ui/components/codeblock';
 import { UsageTabsSelectorLazy, UsageTabLazy } from './lazy';
 import { ResponseTabs } from '../response-tabs';
+import { registerDefault } from '@/requests/generators/all';
 
 /**
  * Generate code example for given programming language
@@ -23,7 +27,7 @@ export interface CodeUsageGenerator<T = unknown> {
    * - a function imported from a file with "use client" directive
    * - false (disabled)
    */
-  source?: string | SampleGenerator<T> | false;
+  source?: string | CodeUsageGeneratorFn<T> | false;
 
   /**
    * Pass extra context to client-side source generator
@@ -50,70 +54,52 @@ export async function UsageTabs({
     );
   };
 
-  renderAPIExampleUsageTabs ??= (generators) => {
-    if (generators.length === 0) return null;
+  renderAPIExampleUsageTabs ??= (registry) => {
+    const map = Array.from(registry.map().entries());
+    if (map.length === 0) return null;
 
     return (
-      <CodeBlockTabs groupId="fumadocs_openapi_requests" defaultValue={generators[0].id}>
+      <CodeBlockTabs groupId="fumadocs_openapi_requests" defaultValue={map[0][0]}>
         <CodeBlockTabsList>
-          {generators.map((item) => (
-            <CodeBlockTabsTrigger key={item.id} value={item.id}>
+          {map.map(([id, item]) => (
+            <CodeBlockTabsTrigger key={id} value={id}>
               {item.label ?? item.lang}
             </CodeBlockTabsTrigger>
           ))}
         </CodeBlockTabsList>
-        {generators.map((item) => (
-          <CodeBlockTab key={item.id} value={item.id}>
-            <UsageTabLazy {...item} />
+        {map.map(([id, item]) => (
+          <CodeBlockTab key={id} value={id}>
+            <UsageTabLazy id={id} lang={item.lang} _client={item._client} />
           </CodeBlockTab>
         ))}
       </CodeBlockTabs>
     );
   };
 
-  let generators: CodeUsageGenerator[] = [...defaultSamples];
-  if (ctx.generateCodeSamples) {
-    generators.push(...(await ctx.generateCodeSamples(method)));
+  let registry: CodeUsageGeneratorRegistry;
+  if (ctx.codeUsages) {
+    registry = createCodeUsageGeneratorRegistry(ctx.codeUsages);
+  } else {
+    registry = createCodeUsageGeneratorRegistry();
+    registerDefault(registry);
+  }
+
+  for (const gen of (await ctx.generateCodeSamples?.(method)) ?? []) {
+    registry.addInline(gen);
   }
 
   if (method['x-codeSamples']) {
     for (const sample of method['x-codeSamples']) {
-      generators.push(
-        'id' in sample && typeof sample.id === 'string'
-          ? (sample as CodeUsageGenerator)
-          : {
-              id: sample.lang,
-              ...sample,
-            },
-      );
+      registry.addInline(sample);
     }
   }
-
-  generators = dedupe(generators);
 
   return renderAPIExampleLayout(
     {
       selector: method['x-exclusiveCodeSample'] ? null : <UsageTabsSelectorLazy />,
-      usageTabs: await renderAPIExampleUsageTabs(generators, ctx),
+      usageTabs: await renderAPIExampleUsageTabs(registry, ctx),
       responseTabs: <ResponseTabs operation={method} ctx={ctx} />,
     },
     ctx,
   );
-}
-
-/**
- * Remove duplicated ids
- */
-function dedupe(samples: CodeUsageGenerator[]): CodeUsageGenerator[] {
-  const set = new Set<string>();
-  const out: CodeUsageGenerator[] = [];
-
-  for (let i = samples.length - 1; i >= 0; i--) {
-    const item = samples[i];
-    if (set.has(item.id)) continue;
-    set.add(item.id);
-    if (item.source !== false) out.unshift(item);
-  }
-
-  return out;
 }